Move sampler back to examples
