--- docs/source/prototype.datasets.rst
@@ -1,11 +1,0 @@
-torchaudio.prototype.datasets
-=============================
-
-.. py:module:: torchaudio.prototype.datasets
-
-.. currentmodule:: torchaudio.prototype.datasets
-
-BucketizeBatchSampler
-~~~~~~~~~~~~~~~~~~~~~
-
-.. autoclass:: BucketizeBatchSampler
