+++ examples/hubert/dataset/__init__.py
@@ -1,10 +1,12 @@
 from .hubert_dataset import (
+    BucketizeBatchSampler,
     CollateFnHubert,
     HuBERTDataSet,
 )
 
 
 __all__ = [
+    "BucketizeBatchSampler",
     "CollateFnHubert",
     "HuBERTDataSet",
 ]
