--- torchaudio/prototype/datasets/__init__.py
@@ -1,5 +1,0 @@
-from .bucketize_sampler import BucketizeBatchSampler
-
-__all__ = [
-    "BucketizeBatchSampler",
-]
